+++ packages/deposit-screener/src/cli/index.ts
@@ -3,9 +3,11 @@
 import { program } from "commander";
 import run from "./commands/run";
 import * as dotenv from "dotenv";
+import { setupDefaultInstrumentation } from "@nocturne-xyz/offchain-utils";
 
 export default async function main(): Promise<void> {
   dotenv.config();
+  setupDefaultInstrumentation("deposit-screener");
 
   program
     .name("screener-cli")