+++ packages/offchain-utils/src/index.ts
@@ -2,3 +2,4 @@ export * from "./loggers";
 export * from "./request";
 export * from "./types";
 export * from "./redis";
+export * from "./instrumentation";